Isolate bottomsheet hook
